--- examples/src/test/java/com/spbsu/flamestream/example/ExampleChecker.java
@@ -1,13 +1,0 @@
-package com.spbsu.flamestream.example;
-
-import java.util.stream.Stream;
-
-/**
- * User: Artem
- * Date: 28.09.2017
- */
-public interface ExampleChecker<T, R> {
-  Stream<T> input();
-
-  void assertCorrect(Stream<R> output);
-}
